Experiments/requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
efficientnet-pytorch==0.6.3
filelock==3.0.12
future==0.18.2
google==3.0.0
google-auth==1.21.2
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==2.0.0
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
msgpack==1.0.0
multidict==4.7.6
munch==2.5.0
networkx==2.5
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
pandas==1.0.4
Pillow==7.2.0
pretrainedmodels==0.7.4
protobuf==3.13.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
ray==0.8.6
redis==3.4.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.2
segmentation-models-pytorch==0.1.0
Shapely==1.7.1
SimpleITK==1.1.0
six==1.15.0
sk-video==1.1.10
soupsieve==2.0.1
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.1.20
torch==1.5.0
torchfile==0.1.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.49.0
typing-extensions==3.7.4.3
urllib3==1.25.10
visdom==0.1.8.9
websocket-client==0.57.0
Werkzeug==1.0.1
yarl==1.5.1
zipp==3.2.0
